Code refactor: Using trait, case classes, & tidy up / reduce file sizes
