+++ src/main/scala/com/siriusxm/example/cart/ShoppingCart.scala
@@ -1,35 +1,19 @@
 package com.siriusxm.example.cart
 
-import zio.UIO
-import zio.Task
-import zio.Ref
-import zio.ZIO
-import com.siriusxm.example.cart.CerealProductInfo.priceLookup
+import com.siriusxm.example.cart.Cart
+import com.siriusxm.example.dto.ProductInfo
+import com.siriusxm.example.service.CerealService.priceLookup
+import zio.{Ref, Task, UIO, ZIO}
 
 import scala.math.BigDecimal
 
-object ShoppingCart:
-  // Tax payable, charged at 12.5% on the subtotal
-  private val TaxRate = 0.125d
-  private val DecimalScale = 2 // money rounded to 2 decimal places
-  private val RoundingMode: BigDecimal.RoundingMode.Value = BigDecimal.RoundingMode.UP
-
-  private type Entries = Map[ProductTitle, (Price, Count)]
-  private type ProductTitle = String
-  private type Price = Float
-  private type Count = Int
-  private val EmptyEntries: Entries = Map.empty
-
-  def newCart: UIO[ShoppingCart] =
-    Ref.make(EmptyEntries).map(new ShoppingCart(_))
-
-  case class ProductInfo(title: String, price: Float)
+object ShoppingCart extends Cart:
 
   class ShoppingCart(private val data: Ref[Entries]):
     /** Add line item by product title, looking up price. */
     def addLineItem(title: String, count: Int): Task[Unit] =
       for price <- priceLookup(title)
-        _ <- addLineItem(ProductInfo(title, price), count)
+          _ <- addLineItem(ProductInfo(title, price), count)
       yield ()
 
     /** Add line item by ProductInfo. */
@@ -61,7 +45,7 @@ object ShoppingCart:
     def totalPayable: UIO[BigDecimal] =
       for
         st <- subtotal
-        tax    <- taxPayable
+        tax <- taxPayable
       yield (st + tax).setScale(DecimalScale, RoundingMode)
 
     /** Number of products in cart */
