+++ src/main/scala/com/siriusxm/example/dto/ProductInfo.scala
@@ -1,0 +1,3 @@
+package com.siriusxm.example.dto
+
+case class ProductInfo(title: String, price: Float)
