+++ src/main/scala/com/siriusxm/example/service/CerealService.scala
@@ -1,13 +1,13 @@
-package com.siriusxm.example.cart
+package com.siriusxm.example.service
 
-import com.siriusxm.example.cart.ShoppingCart.ProductInfo
+import com.siriusxm.example.dto.ProductInfo
 import sttp.client4.httpclient.zio.HttpClientZioBackend
 import sttp.client4.ziojson.asJson
 import sttp.client4.{UriContext, basicRequest}
 import zio.ZIO
 import zio.json.{DeriveJsonDecoder, JsonDecoder}
 
-object CerealProductInfo:
+object CerealService:
   private val baseUrl = "https://raw.githubusercontent.com/mattjanks16/shopping-cart-test-data/main"
 
   given JsonDecoder[ProductInfo] = DeriveJsonDecoder.gen[ProductInfo]
@@ -22,8 +22,7 @@ object CerealProductInfo:
         for {
           backend <- HttpClientZioBackend() // Create the HTTP client
           response <- request.response(asJson[ProductInfo]).send(backend) // Send the request
-          _ <- ZIO.logInfo(s"Response Code: ${response.code}")
-          _ <- ZIO.logInfo(s"Response Body: ${response.body}")
+          _ <- ZIO.logInfo(s"Response Code: ${response.code}, Response Body: ${response.body}")
           ret <- ZIO.fromEither(response.body.map(_.price)).orElse(ZIO.from(0.0f))// return price or default to 0.0
         } yield ret
       }
